Render Kick chats

+++ TwitchDownloaderCore/ChatUpdater.cs
@@ -173,8 +173,8 @@ namespace TwitchDownloaderCore
                 newEmote.id = emote.Id;
                 newEmote.imageScale = emote.ImageScale;
                 newEmote.data = emote.ImageData;
-                newEmote.width = emote.Width / emote.ImageScale;
-                newEmote.height = emote.Height / emote.ImageScale;
+                newEmote.width = (int)(emote.Width / emote.ImageScale);
+                newEmote.height = (int)(emote.Height / emote.ImageScale);
                 chatRoot.embeddedData.firstParty.Add(newEmote);
             }
             progress?.Report(new ProgressReport(ReportType.Log, $"Input 1st party emote count: {inputCount}. Output count: {chatRoot.embeddedData.firstParty.Count}"));
@@ -193,8 +193,8 @@ namespace TwitchDownloaderCore
                 newEmote.imageScale = emote.ImageScale;
                 newEmote.data = emote.ImageData;
                 newEmote.name = emote.Name;
-                newEmote.width = emote.Width / emote.ImageScale;
-                newEmote.height = emote.Height / emote.ImageScale;
+                newEmote.width = (int)(emote.Width / emote.ImageScale);
+                newEmote.height = (int)(emote.Height / emote.ImageScale);
                 chatRoot.embeddedData.thirdParty.Add(newEmote);
             }
             progress?.Report(new ProgressReport(ReportType.Log, $"Input 3rd party emote count: {inputCount}. Output count: {chatRoot.embeddedData.thirdParty.Count}"));
@@ -234,8 +234,8 @@ namespace TwitchDownloaderCore
                     newEmote.imageScale = emotePair.Value.ImageScale;
                     newEmote.data = emotePair.Value.ImageData;
                     newEmote.name = emotePair.Value.Name;
-                    newEmote.width = emotePair.Value.Width / emotePair.Value.ImageScale;
-                    newEmote.height = emotePair.Value.Height / emotePair.Value.ImageScale;
+                    newEmote.width = (int)(emotePair.Value.Width / emotePair.Value.ImageScale);
+                    newEmote.height = (int)(emotePair.Value.Height / emotePair.Value.ImageScale);
                     newBit.tierList.Add(emotePair.Key, newEmote);
                 }
                 chatRoot.embeddedData.twitchBits.Add(newBit);
